requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dataclasses-json==0.5.8
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.2
django-import-export==3.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
frozenlist==1.3.3
idna==3.4
kombu==5.3.1
magic-filter==1.0.9
MarkupPy==1.14
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
Pillow==9.5.0
prompt-toolkit==3.0.38
pydantic==1.10.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-interface==1.6.1
pytz==2023.3
PyYAML==6.0
redis==4.5.5
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.2
